refactor(nika-cap): builtin arg-shape rules move in — schema back under its cap
The E1×E2 union put nika-schema at 15277/15000 prod LOC (it sat at
exactly 15000 before this arc — zero headroom). Operator-approved
extraction, the 2026-07-03 permits-half precedent: the builtin
arg-shape rules (done sentinel · wait XOR · the whole fetch contract
incl. the vNext payload/traverse families · image/tts enums+ranges)
become nika_cap::builtin_shape_findings — PURE DATA over
serde_json::Value, messages out, no parser types, no spans.

nika-schema keeps a ~45-line adapter (task walk + span wrapping into
SchemaError::BadBuiltinArgs) AND its full truth-table test corpus —
every table still runs yaml → analyze → cap rules → diagnostics, so
the two crates cannot drift. nika-schema: 15277 → 14650 · nika-cap:
810 → 1250 (both far under cap).

Gates: 21 cap + 779 schema lib (39/39 workspace) · clippy 0
all-targets · crate-size OK · public-api cap +1 (local regen — pure
data crate, no platform cfg; CI arbitrates).
