+++ crates/nika-cap/Cargo.toml
@@ -12,6 +12,7 @@ publish                = false  # Foundation crate — never on crates.io. See A
 
 [dependencies]
 nika-types = { path = "../nika-types", version = "0.97.0" }  # host_glob_matches — the same matcher nika-http enforces at runtime (zero drift)
+serde_json = { workspace = true }   # builtin arg-shape rules — pure data over Value (extracted 2026-07-07)
 serde      = { workspace = true }
 
 [dev-dependencies]
