+++ crates/nika-schema/src/analyzer/builtin_shape.rs
@@ -1,19 +1,21 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 // Copyright (C) 2024-2026 SuperNovae Studio <contact@supernovae.studio>
 
-//! Builtin arg-shape rules — the statically-checkable contracts of
-//! `stdlib/builtins-v0.1.md` the JSON Schema cannot express (deep
-//! conformance fixtures 009-012) · `nika:write` requires `content:` ·
-//! `nika:done` is valid only inside an `agent:` tools whitelist (the
-//! loop sentinel · `NIKA-BUILTIN-DONE-001`) · `nika:jq` takes exactly
-//! `expression:` · `nika:wait` takes `duration:` XOR `until:`.
-
-use nika_types::extract::{EXTRACT_MODE_NAMES, ExtractMode};
-use nika_types::net::MAX_TRAVERSE_PAGES;
+//! Builtin arg-shape rules — the thin analyzer adapter.
+//!
+//! The RULES live in `nika-cap::builtin_shape_findings` (pure data ·
+//! extracted 2026-07-07 when this crate hit its 15k prod budget — the
+//! permits-half precedent of 2026-07-03): this side only walks the
+//! parsed tasks (main action + `on_finally:` cleanups), hands each
+//! `invoke:`'s (tool · args) to the shared rule set, and wraps every
+//! finding message in [`SchemaError::BadBuiltinArgs`] with the task's
+//! span. The truth-table tests below exercise the WHOLE pipe
+//! (yaml → analyze → cap rules → diagnostics), so the two crates
+//! cannot drift.
 
 use crate::error::SchemaError;
 use crate::raw::{RawAction, RawTask};
-use crate::source::{Span, Spanned};
+use crate::source::Spanned;
 
 /// Run every builtin arg-shape rule over a task's action (and its
 /// `on_finally:` cleanup actions — same `invoke:` surface).
@@ -34,668 +36,13 @@ fn check_action(action: &RawAction, task: &str, errors: &mut Vec<SchemaError>) {
     let tool = invoke.tool.value.as_str();
     let span = invoke.tool.span;
     let args = invoke.args.as_ref().map(|a| &a.value);
-    let has = |key: &str| -> bool {
-        matches!(args, Some(serde_json::Value::Object(map)) if map.contains_key(key))
-    };
-
-    // Flat required-arg contracts (`nika:write` content · `nika:jq`
-    // expression · `nika:fetch` url · …) now live in the catalog's
-    // `Builtin::required` and are checked by `check::tools::scan_missing_args`
-    // — ONE source, no double report. This ladder keeps only the
-    // NON-FLAT contracts the catalog set cannot express.
-    match tool {
-        "nika:done" => errors.push(shape(
-            task,
-            tool,
-            "is the agent-loop completion sentinel — valid ONLY inside an \
-             `agent:` tools whitelist · never a standalone invoke \
-             (02-verbs.md §loop semantics · NIKA-BUILTIN-DONE-001)",
-            span,
-        )),
-        "nika:wait" if has("duration") == has("until") => errors.push(shape(
-            task,
-            tool,
-            "takes `duration:` XOR `until:` — exactly one \
-             (builtins-v0.1.md §nika:wait)",
-            span,
-        )),
-        // `nika:fetch` keeps its WHOLE contract here (incl. the required
-        // `url:`) because the mode-dependent `jq`/`selector` pairings are
-        // conditional — splitting `url` to the catalog set would double the
-        // surface for one builtin. Its `Builtin::required` stays empty.
-        "nika:fetch" => check_fetch_shape(task, tool, args, span, errors),
-        "nika:image_generate" => check_image_generate_shape(task, tool, args, span, errors),
-        "nika:tts_generate" => check_tts_generate_shape(task, tool, args, span, errors),
-        _ => {}
-    }
-}
-
-/// `nika:tts_generate` static contracts (stdlib §Audio): closed enums +
-/// literal numeric ranges, mirroring the runtime parser — templated
-/// values are runtime business (statically silent).
-fn check_tts_generate_shape(
-    task: &str,
-    tool: &str,
-    args: Option<&serde_json::Value>,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    let object = args.and_then(serde_json::Value::as_object);
-    let literal = |key: &str| -> Option<&str> {
-        let value = object?.get(key)?.as_str()?;
-        (!value.contains("${{")).then_some(value)
-    };
-
-    let enums: [(&str, &[&str]); 2] = [
-        ("provider", &["local", "openai", "elevenlabs", "mock"]),
-        ("format", &["mp3", "wav", "auto"]),
-    ];
-    for (key, allowed) in enums {
-        if let Some(value) = literal(key)
-            && !allowed.contains(&value)
-        {
-            errors.push(shape(
-                task,
-                tool,
-                &format!(
-                    "`{key}: {value}` is not in the closed set — {}",
-                    allowed.join(" · ")
-                ),
-                span,
-            ));
-        }
-    }
-    if let Some(value) = object.and_then(|map| map.get("speed"))
-        && let Some(s) = value.as_f64()
-        && !(0.25..=4.0).contains(&s)
-    {
-        errors.push(shape(
-            task,
-            tool,
-            &format!("`speed: {s}` out of range — 0.25..=4.0"),
-            span,
-        ));
-    }
-    if let Some(value) = object.and_then(|map| map.get("timeout_ms"))
-        && let Some(n) = value.as_i64()
-        && !(1_000..=600_000).contains(&n)
-    {
-        errors.push(shape(
-            task,
-            tool,
-            &format!("`timeout_ms: {n}` out of range — 1000..=600000"),
-            span,
-        ));
-    }
-}
-
-/// `nika:fetch` static contracts (stdlib §fetch + extract-modes-v0.1.md ·
-/// conformance `stdlib/extract-modes/001..004`): the mode set is CLOSED
-/// at v0.1 · `jq:` pairs with `mode: jq` exactly · `selector:` pairs
-/// with `mode: selector` exactly. A templated `mode:` (`${{ … }}`) is
-/// runtime business — statically silent.
-fn check_fetch_shape(
-    task: &str,
-    tool: &str,
-    args: Option<&serde_json::Value>,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    let object = args.and_then(serde_json::Value::as_object);
-
-    // `url:` is fetch's one REQUIRED argument — a fetch with nothing to
-    // fetch should die at check time, not at runtime (review lens 1 P1).
-    if !object.is_some_and(|map| map.contains_key("url")) {
-        errors.push(shape(
-            task,
-            tool,
-            "requires a `url:` argument — a fetch with nothing to fetch \
-             (builtins-v0.1.md §nika:fetch)",
-            span,
-        ));
-        return;
-    }
-    check_fetch_payload_shape(task, tool, object, span, errors);
-    if object.is_some_and(|map| map.contains_key("traverse")) {
-        check_fetch_traverse_shape(task, tool, object, span, errors);
-        return; // traverse owns the whole surface — no mode pairing below
-    }
-    let mode = match object.and_then(|map| map.get("mode")) {
-        // Absent → the spec default (markdown).
-        None => Some(ExtractMode::Markdown),
-        Some(serde_json::Value::String(raw)) => {
-            if raw.contains("${{") {
-                None // templated — the pairing is unknowable statically
-            } else if let Ok(parsed) = raw.parse::<ExtractMode>() {
-                Some(parsed)
-            } else {
-                errors.push(shape(
-                    task,
-                    tool,
-                    &format!(
-                        "`mode: {raw}` is not a stdlib v0.1 extract mode — the set \
-                         is closed: {EXTRACT_MODE_NAMES} (extract-modes-v0.1.md)"
-                    ),
-                    span,
-                ));
-                return;
-            }
-        }
-        Some(_) => {
-            errors.push(shape(
-                task,
-                tool,
-                "`mode:` must be a string (extract-modes-v0.1.md)",
-                span,
-            ));
-            return;
-        }
-    };
-
-    let has = |key: &str| object.is_some_and(|map| map.contains_key(key));
-    let Some(mode) = mode else { return };
-
-    if has("jq") && mode != ExtractMode::Jq {
-        errors.push(shape(
-            task,
-            tool,
-            "`jq:` is «a jq expression · only with mode: jq» — pair it with \
-             `mode: jq` or drop it (builtins-v0.1.md §nika:fetch)",
-            span,
-        ));
-    }
-    if mode == ExtractMode::Jq && !has("jq") {
-        errors.push(shape(
-            task,
-            tool,
-            "`mode: jq` requires the `jq:` expression to apply \
-             (builtins-v0.1.md §nika:fetch)",
-            span,
-        ));
-    }
-    if has("selector") && mode != ExtractMode::Selector {
-        errors.push(shape(
-            task,
-            tool,
-            "`selector:` pairs with `mode: selector` only \
-             (extract-modes-v0.1.md §selector)",
-            span,
-        ));
-    }
-    if mode == ExtractMode::Selector && !has("selector") {
-        errors.push(shape(
-            task,
-            tool,
-            "`mode: selector` requires the `selector:` CSS selector \
-             (extract-modes-v0.1.md §selector)",
-            span,
-        ));
-    }
-}
-
-/// `traverse:` static contracts (stdlib §fetch · traverse): the crawl
-/// excludes the single-fetch extraction args (`mode`/`selector`/`jq` —
-/// the payload families are already covered by the exclusivity rule in
-/// [`check_fetch_payload_shape`] since traverse is GET-only), forces
-/// GET, and its own shape is CLOSED (`max_pages` 1..=25 required ·
-/// `respect_robots` bool). Templated values are runtime business.
-fn check_fetch_traverse_shape(
-    task: &str,
-    tool: &str,
-    object: Option<&serde_json::Map<String, serde_json::Value>>,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    let has = |key: &str| object.is_some_and(|map| map.contains_key(key));
-    for key in ["mode", "selector", "jq", "body", "form", "multipart"] {
-        if has(key) {
-            errors.push(shape(
-                task,
-                tool,
-                &format!(
-                    "`traverse:` excludes `{key}:` — the crawl emits the fixed \
-                     page-digest shape (builtins-v0.1.md §nika:fetch · traverse)"
-                ),
-                span,
-            ));
-        }
-    }
-    if let Some(method) = object
-        .and_then(|map| map.get("method"))
-        .and_then(serde_json::Value::as_str)
-        && !method.contains("${{")
-        && !method.eq_ignore_ascii_case("GET")
-    {
-        errors.push(shape(
-            task,
-            tool,
-            &format!("`traverse:` crawls with GET only — drop `method: {method}`"),
-            span,
-        ));
-    }
-    let traverse = object.and_then(|map| map.get("traverse"));
-    let map = match traverse {
-        Some(serde_json::Value::String(s)) if s.contains("${{") => return, // runtime
-        Some(serde_json::Value::Object(map)) => map,
-        Some(_) | None => {
-            errors.push(shape(
-                task,
-                tool,
-                "`traverse:` must be an object — `{ max_pages: N, respect_robots?: bool }`",
-                span,
-            ));
-            return;
-        }
-    };
-    if let Some(unknown) = map
-        .keys()
-        .find(|k| !matches!(k.as_str(), "max_pages" | "respect_robots"))
-    {
-        errors.push(shape(
-            task,
-            tool,
-            &format!("`traverse.{unknown}:` is not a traverse field — the shape is closed"),
-            span,
-        ));
-    }
-    match map.get("max_pages") {
-        None => errors.push(shape(
-            task,
-            tool,
-            &format!(
-                "`traverse.max_pages:` is required — an integer 1..={MAX_TRAVERSE_PAGES} \
-                 (the crawl bound)"
-            ),
-            span,
-        )),
-        Some(serde_json::Value::String(s)) if s.contains("${{") => {} // runtime
-        Some(value) => match value.as_u64() {
-            Some(n) if (1..=MAX_TRAVERSE_PAGES).contains(&n) => {}
-            Some(n) => errors.push(shape(
-                task,
-                tool,
-                &format!("`traverse.max_pages: {n}` out of range — 1..={MAX_TRAVERSE_PAGES}"),
-                span,
-            )),
-            None => errors.push(shape(
-                task,
-                tool,
-                &format!("`traverse.max_pages:` must be an integer 1..={MAX_TRAVERSE_PAGES}"),
-                span,
-            )),
-        },
-    }
-    match map.get("respect_robots") {
-        None | Some(serde_json::Value::Bool(_)) => {}
-        Some(serde_json::Value::String(s)) if s.contains("${{") => {} // runtime
-        Some(_) => errors.push(shape(
-            task,
-            tool,
-            "`traverse.respect_robots:` must be a boolean",
-            span,
-        )),
-    }
-}
-
-/// The fetch vNext payload families (stdlib §fetch): `body ⊥ form ⊥
-/// multipart` · `form:`/`multipart:` need a body-bearing method · the
-/// multipart part shape is CLOSED. A templated value (`${{ … }}`) is
-/// runtime business — statically silent (the runtime re-vets all of it).
-fn check_fetch_payload_shape(
-    task: &str,
-    tool: &str,
-    object: Option<&serde_json::Map<String, serde_json::Value>>,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    let has = |key: &str| object.is_some_and(|map| map.contains_key(key));
-    if ["body", "form", "multipart"]
-        .iter()
-        .filter(|key| has(key))
-        .count()
-        > 1
-    {
-        errors.push(shape(
-            task,
-            tool,
-            "at most one of `body:` · `form:` · `multipart:` \
-             (builtins-v0.1.md §nika:fetch)",
-            span,
-        ));
-    }
-    if (has("form") || has("multipart"))
-        && let Some(headers) = object
-            .and_then(|map| map.get("headers"))
-            .and_then(serde_json::Value::as_object)
-        && headers
-            .keys()
-            .any(|k| k.eq_ignore_ascii_case("content-type"))
-    {
-        errors.push(shape(
-            task,
-            tool,
-            "`form:`/`multipart:` set their own content-type — drop the \
-             `headers:` entry (builtins-v0.1.md §nika:fetch)",
-            span,
-        ));
-    }
-    if has("form") || has("multipart") {
-        match object
-            .and_then(|map| map.get("method"))
-            .and_then(serde_json::Value::as_str)
-        {
-            None => errors.push(shape(
-                task,
-                tool,
-                "`form:`/`multipart:` need `method: POST` (or PUT/PATCH) — \
-                 the default GET carries no body",
-                span,
-            )),
-            Some(raw) if raw.contains("${{") => {} // templated — runtime business
-            Some(raw) => {
-                let upper = raw.to_uppercase();
-                if matches!(upper.as_str(), "GET" | "HEAD" | "DELETE") {
-                    errors.push(shape(
-                        task,
-                        tool,
-                        &format!(
-                            "`form:`/`multipart:` need a body-bearing method — \
-                             `{raw}` carries no body (use POST · PUT · PATCH)"
-                        ),
-                        span,
-                    ));
-                }
-            }
-        }
-    }
-    if let Some(form) = object.and_then(|map| map.get("form"))
-        && !form.is_object()
-        && !matches!(form, serde_json::Value::String(s) if s.contains("${{"))
-    {
-        errors.push(shape(
-            task,
-            tool,
-            "`form:` must be an object of scalar fields (builtins-v0.1.md §nika:fetch)",
-            span,
-        ));
-    }
-    if let Some(parts) = object.and_then(|map| map.get("multipart")) {
-        check_multipart_parts(task, tool, parts, span, errors);
-    }
-}
-
-/// The closed multipart part shape — `{name, value}` XOR `{name, path,
-/// filename?, content_type?}`. Presence rules are static even when the
-/// VALUES are templated; a fully-templated `multipart:` string is silent.
-fn check_multipart_parts(
-    task: &str,
-    tool: &str,
-    parts: &serde_json::Value,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    use nika_types::net::MULTIPART_PART_KEYS as PART_KEYS;
-    let items = match parts {
-        serde_json::Value::String(s) if s.contains("${{") => return,
-        serde_json::Value::Array(items) => items,
-        _ => {
-            errors.push(shape(
-                task,
-                tool,
-                "`multipart:` must be an array of parts (builtins-v0.1.md §nika:fetch)",
-                span,
-            ));
-            return;
-        }
-    };
-    if items.is_empty() {
-        errors.push(shape(
-            task,
-            tool,
-            "`multipart:` needs at least one part",
-            span,
-        ));
-        return;
-    }
-    for (i, item) in items.iter().enumerate() {
-        let Some(map) = item.as_object() else {
-            errors.push(shape(
-                task,
-                tool,
-                &format!("multipart part {i} must be an object"),
-                span,
-            ));
-            continue;
-        };
-        if let Some(unknown) = map.keys().find(|k| !PART_KEYS.contains(&k.as_str())) {
-            errors.push(shape(
-                task,
-                tool,
-                &format!(
-                    "multipart part {i}: unknown key `{unknown}` — the shape is \
-                     {{name, value}} or {{name, path, filename?, content_type?}}"
-                ),
-                span,
-            ));
-        }
-        if !map.contains_key("name") {
-            errors.push(shape(
-                task,
-                tool,
-                &format!("multipart part {i} needs a `name:`"),
-                span,
-            ));
-        }
-        match (map.contains_key("value"), map.contains_key("path")) {
-            (true, true) | (false, false) => errors.push(shape(
-                task,
-                tool,
-                &format!("multipart part {i}: exactly one of `value:` (text) | `path:` (file)"),
-                span,
-            )),
-            (true, false) => {
-                if map.contains_key("filename") || map.contains_key("content_type") {
-                    errors.push(shape(
-                        task,
-                        tool,
-                        &format!(
-                            "multipart part {i}: `filename:`/`content_type:` belong to \
-                             file parts (`path:`)"
-                        ),
-                        span,
-                    ));
-                }
-            }
-            (false, true) => {}
-        }
-    }
-}
-
-/// `nika:image_generate` static contracts (stdlib §Media): the V1
-/// reservations are refused loudly, the closed literal enums + numeric
-/// ranges mirror the runtime parser (`nika-builtin::image::args` — same
-/// values, caught at check time instead of after a spent request), and
-/// the transparent-background × jpeg conflict dies here too. A templated
-/// value (`${{ … }}`) is runtime business — statically silent. The flat
-/// required args (`prompt:` · `output_dir:`) live in the catalog
-/// `Builtin::required` set (checked by `check::tools::scan_missing_args`).
-fn check_image_generate_shape(
-    task: &str,
-    tool: &str,
-    args: Option<&serde_json::Value>,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    let object = args.and_then(serde_json::Value::as_object);
-    let literal = |key: &str| -> Option<&str> {
-        let value = object?.get(key)?.as_str()?;
-        (!value.contains("${{")).then_some(value)
-    };
-
-    check_image_v1_reservations(task, tool, object, span, errors);
-
-    // Closed literal enums (builtins-v0.1.md §nika:image_generate).
-    let enums: [(&str, &[&str]); 5] = [
-        ("provider", &["local", "openai", "gemini", "xai", "mock"]),
-        ("format", &["png", "jpeg", "jpg", "webp"]),
-        ("quality", &["auto", "low", "medium", "high", "ultra"]),
-        ("background", &["auto", "transparent", "opaque"]),
-        (
-            "aspect_ratio",
-            &["1:1", "16:9", "9:16", "4:3", "3:4", "3:2", "2:3", "21:9"],
-        ),
-    ];
-    for (key, allowed) in enums {
-        if let Some(value) = literal(key)
-            && !allowed.contains(&value)
-        {
-            errors.push(shape(
-                task,
-                tool,
-                &format!(
-                    "`{key}: {value}` is not in the closed set — {}",
-                    allowed.join(" · ")
-                ),
-                span,
-            ));
-        }
-    }
-
-    // Literal numeric ranges (same bounds as the runtime parser).
-    let ranges: [(&str, i64, i64); 3] = [
-        ("n", 1, 10),
-        ("compression", 0, 100),
-        ("timeout_ms", 1_000, 600_000),
-    ];
-    for (key, min, max) in ranges {
-        if let Some(value) = object.and_then(|map| map.get(key))
-            && let Some(n) = value.as_i64()
-            && !(min..=max).contains(&n)
-        {
-            errors.push(shape(
-                task,
-                tool,
-                &format!("`{key}: {n}` is out of range — {min}..={max}"),
-                span,
-            ));
-        }
-    }
-
-    // `size:` grammar — `auto` or `WIDTHxHEIGHT`.
-    if let Some(value) = literal("size")
-        && value != "auto"
-        && !value
-            .split_once('x')
-            .is_some_and(|(w, h)| w.parse::<u32>().is_ok() && h.parse::<u32>().is_ok())
-    {
-        errors.push(shape(
-            task,
-            tool,
-            &format!("`size: {value}` must be `WIDTHxHEIGHT` (e.g. 1024x1024) or `auto`"),
-            span,
-        ));
-    }
-
-    // Transparency needs an alpha-capable format — the one cross-arg
-    // conflict decidable statically (provider/model support is runtime).
-    if literal("background") == Some("transparent")
-        && matches!(literal("format"), Some("jpeg" | "jpg"))
-    {
-        errors.push(shape(
-            task,
-            tool,
-            "`background: transparent` needs an alpha-capable `format:` — png or webp \
-             (builtins-v0.1.md §nika:image_generate)",
-            span,
-        ));
-    }
-}
-
-/// The V1 reservations — structurally parsed, loudly refused, never
-/// silently ignored (the mission's honest-boundary contract).
-fn check_image_v1_reservations(
-    task: &str,
-    tool: &str,
-    object: Option<&serde_json::Map<String, serde_json::Value>>,
-    span: Span,
-    errors: &mut Vec<SchemaError>,
-) {
-    let literal = |key: &str| -> Option<&str> {
-        let value = object?.get(key)?.as_str()?;
-        (!value.contains("${{")).then_some(value)
-    };
-    let is_edit = match literal("mode") {
-        None | Some("generate") => false,
-        Some("edit") => true,
-        Some(other) => {
-            errors.push(shape(
-                task,
-                tool,
-                &format!("`mode: {other}` is not a mode — one of generate · edit"),
-                span,
-            ));
-            false
-        }
-    };
-    let has = |k: &str| object.is_some_and(|m| m.contains_key(k));
-    if is_edit {
-        // edit REQUIRES a source; `image` XOR `images`; a templated path is
-        // runtime business (statically silent).
-        if !has("image") && !has("images") {
-            errors.push(shape(
-                task,
-                tool,
-                "`mode: edit` requires `image:` (a path) or `images:` (paths)",
-                span,
-            ));
-        }
-        if has("image") && has("images") {
-            errors.push(shape(
-                task,
-                tool,
-                "`image:` and `images:` are mutually exclusive — use one",
-                span,
-            ));
-        }
-    } else {
-        // edit-only keys in generate mode are a loud static error.
-        for key in ["image", "images", "mask"] {
-            if has(key) {
-                errors.push(shape(
-                    task,
-                    tool,
-                    &format!("`{key}:` requires `mode: edit`"),
-                    span,
-                ));
-            }
-        }
-    }
-    if object.is_some_and(|map| map.contains_key("reference_images")) {
-        errors.push(shape(
-            task,
-            tool,
-            "`reference_images:` is not in V1 — text-to-image only; reference-image \
-             composition is on the media roadmap",
-            span,
-        ));
-    }
-    if object.and_then(|map| map.get("save")) == Some(&serde_json::Value::Bool(false)) {
-        errors.push(shape(
-            task,
-            tool,
-            "`save: false` is not in V1 — assets always land in `output_dir:` \
-             (image bytes never ride workflow outputs)",
-            span,
-        ));
-    }
-}
-
-fn shape(task: &str, tool: &str, reason: &str, span: Span) -> SchemaError {
-    SchemaError::BadBuiltinArgs {
-        task: task.to_owned(),
-        tool: tool.to_owned(),
-        reason: reason.to_owned(),
-        span: Some(span),
+    for reason in nika_cap::builtin_shape_findings(tool, args) {
+        errors.push(SchemaError::BadBuiltinArgs {
+            task: task.to_owned(),
+            tool: tool.to_owned(),
+            reason,
+            span: Some(span),
+        });
     }
 }
